requirements.txt
# proj
celery==5.2.7
colorlog==6.7.0
Django==4.1.7
django-celery-results==2.4.0
django-environ==0.10.0
djangorestframework==3.14.0
gevent==22.10.2